requirements.txt
beautifulsoup4
requests
pandas
pillow
aiohttp
coloredlogs
rich-click
rich